requirements.txt
mysql-replication==1.0.5
boto3==1.34.9
cryptography
pyiceberg[glue,s3fs,pyarrow]
pydantic>=2.0.0
